PREP 3:
Finished the core requirements, initially the program asked a user for a "magic number" which they then tried to guess.
This was later replaced with a random number generator that gets a number between 1 and 100.
The program uses a do-while loop to keep the user playing until they guess the magic number.