W03 Learning02 Assignment:
Created a new class file (Job.cs) and a new class file (Resume.cs) in the project.
The Job class has variables for the job title, company name, start date, and end date. It also has a method to display them when called.
The Resume class has variables for the name and a list of jobs. It also has a method to display them when called.