Develop02:
Finished the requirements for the program:

-Write a new entry - Show the user a random prompt (from a list that you create), and save their response, the prompt, and the date as an Entry.
-Display the journal - Iterate through all entries in the journal and display them to the screen.
-Save the journal to a file - Prompt the user for a filename and then save the current journal (the complete list of entries) to that file location.
-Load the journal from a file - Prompt the user for a filename and then load the journal (a complete list of entries) from that file. This should replace any entries currently stored in the journal.
-Provide a menu that allows the user choose these options
-Your list of prompts must contain at least five different prompts. Make sure to add your own prompts to the list, but the following are examples to help get you started:

Journal prompts are stored in their own class and chosen in that at random.
Journal class is created to store the entries and handle the saving and loading of the journal.
Program class is updated to handle the new requirements and the menu is created to allow the user to choose the options.
Entry Class handles the actual data and formatting of the entry and prompts for the file and journal class.