Develop03:
Files added:
- ScriptureList.cs
- reference.cs
- scripture.cs
- word.cs

Program has some set scriptures for the user to choose from. Future updates to the program will allow the user to add their own scriptures to it and set everything in a file instead of in the program.
To help with memorization, the user through the enter key can have words removed (currenlty 3 at a time) as they work through the verse.